+++ finagle-thrift/src/main/scala/com/twitter/finagle/thrift/service/Filterable.scala
@@ -6,7 +6,7 @@ import com.twitter.finagle.Filter
  * Used in conjunction with a `ServicePerEndpoint` builder to allow for filtering
  * of a `ServicePerEndpoint`.
  */
-trait Filterable[+T] {
+trait Filterable[+T] extends com.twitter.finagle.service.Filterable[T] {
 
   /**
    * Prepend the given type-agnostic [[Filter]].